Build the training model
In order to have a model that will help differentiate certain facial expressions,
we must first train the model. To do this, I'll be using the PyTorch library that will come
with some useful functions. The "training-model.py" program won't be given a dataset just yet
to train but I'll soon give it later down the line.