Finalize the project
I've updated program main.py to include the model "emotion_model.pt." The program will open a camera to detect
faces and display a text in the frame suggesting the emotion the model thinks the face is making.
I've also added emotion_model.pt for use, losses.png to take a look at the loss plot.